Fix remaining security vulnerability in react-router
- Added overrides for `react-router` and `react-router-dom` to `^7.12.0` in `package.json`.
- Updated `package-lock.json` via `npm install`.
- Verified that `react-router` and `react-router-dom` are updated to version 7.13.0 (which satisfies ^7.12.0) using `npm list`.
- All tests passed.
