+++ npm_install_sec_2.log
@@ -1,0 +1,3 @@
+npm error code EOVERRIDE
+npm error Override for react-router-dom@^7.11.0 conflicts with direct dependency
+npm error A complete log of this run can be found in: /home/jules/.npm/_logs/2026-02-10T01_52_50_545Z-debug-0.log
