+++ package.json
@@ -27,7 +27,7 @@
     "jsonwebtoken": "^9.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^7.11.0",
+    "react-router-dom": "^7.12.0",
     "react-scripts": "^5.0.1",
     "styled-components": "^6.1.1"
   },
@@ -63,6 +63,8 @@
     "@modelcontextprotocol/sdk": "^1.26.0",
     "fast-xml-parser": "^5.3.4",
     "hono": "^4.11.4",
-    "tar": "^7.5.7"
+    "tar": "^7.5.7",
+    "react-router": "^7.12.0",
+    "react-router-dom": "^7.12.0"
   }
 }
