+++ src/nlp/cd-detector.test.js
@@ -1,98 +1,83 @@
+let detectCDs;
 
-const mockEmbedFunction = jest.fn(() => {
-    // Return a tensor-like object synchronously for await to resolve with
-    return Promise.resolve({
-        squeeze: jest.fn(() => ({
-            dispose: jest.fn(),
-            dataSync: jest.fn(() => [0.1]),
-        })),
-        dispose: jest.fn(),
-        dataSync: jest.fn(() => [0.1]),
-    });
+jest.mock('@tensorflow-models/universal-sentence-encoder', () => ({
+  load: jest.fn(),
+}));
+
+const createMockSqueezedTensor = (similarityValue = 0.9) => ({
+  neg: jest.fn(() => createMockSqueezedTensor(similarityValue)),
+  add: jest.fn(() => createMockSqueezedTensor(similarityValue)),
+  dataSync: jest.fn(() => [similarityValue]),
+  dispose: jest.fn(),
 });
 
-jest.mock('@tensorflow-models/universal-sentence-encoder', () => {
+const mockTensor = {
+  squeeze: jest.fn(() => createMockSqueezedTensor()),
+  dispose: jest.fn(),
+};
+
+jest.mock('@tensorflow/tfjs', () => {
+  const originalTf = jest.requireActual('@tensorflow/tfjs');
   return {
-    load: jest.fn(() => Promise.resolve({
-      embed: mockEmbedFunction,
+    ...originalTf,
+    mean: jest.fn((tensor) => mockTensor), // Always return mockTensor
+    metrics: {
+      ...originalTf.metrics,
+      cosineDistance: jest.fn((vec1, vec2) => createMockSqueezedTensor(0.9)), // Default high similarity
+    },
+    tensor2d: jest.fn(() => mockTensor),
+    equal: jest.fn(() => ({
+      all: jest.fn(() => ({
+        dataSync: jest.fn(() => [1]),
+      })),
     })),
   };
 });
 
-const mockTensor = {
-  dataSync: jest.fn(() => [0.9]),
-  dispose: jest.fn(),
-  squeeze: jest.fn(() => mockTensor),
-};
-
-jest.mock('@tensorflow/tfjs', () => {
-  return {
-    tidy: jest.fn((callback) => callback()),
-    mul: jest.fn(() => mockTensor),
-    sum: jest.fn(() => mockTensor),
-    mean: jest.fn(() => mockTensor),
-    tensor2d: jest.fn(() => mockTensor),
-  };
-});
-
-let detectCDs;
-let tf;
-
 describe('detectCDs', () => {
+  let mockEmbed;
+  let use;
+  let tf;
+
   beforeEach(() => {
     jest.resetModules();
+    ({ detectCDs } = require('./cd-detector'));
+    use = require('@tensorflow-models/universal-sentence-encoder');
     tf = require('@tensorflow/tfjs');
 
-    // Explicitly set up the mock return value BEFORE requiring the module
-    // This is critical because detectCDs calls loadModel() on initialization (if not test env, but here we invoke it via detectCDs)
-    // Actually, detectCDs calls loadModel internally if model is null.
+    mockEmbed = jest.fn(() => Promise.resolve(mockTensor));
+    use.load.mockResolvedValue({ embed: mockEmbed });
+  });
 
-    // Reset mocks
-    mockEmbedFunction.mockClear();
-    tf.sum.mockClear();
-
-    ({ detectCDs } = require('./cd-detector'));
+  afterEach(() => {
+    jest.clearAllMocks();
   });
 
   test('should return an empty array if no cognitive distortions are detected', async () => {
-    // Mock sum to return a tensor with low similarity
-    const lowSimTensor = {
-      dataSync: jest.fn(() => [0.1]),
-      dispose: jest.fn(),
-    };
-    tf.sum.mockReturnValue(lowSimTensor);
-
-    // Ensure embed returns a promise that resolves to an object with squeeze
-    mockEmbedFunction.mockResolvedValue({
-        squeeze: jest.fn(() => ({ dispose: jest.fn() })),
-        dispose: jest.fn(),
-        dataSync: jest.fn(() => [0.1]),
-    });
-
+    // Simulate low similarity for all CDs
+    tf.metrics.cosineDistance.mockReturnValue(createMockSqueezedTensor(0.1));
     const text = 'This is a neutral sentence.';
     const result = await detectCDs(text);
-
     expect(result).toEqual([]);
+    expect(use.load).toHaveBeenCalledTimes(1);
+    expect(mockEmbed).toHaveBeenCalled();
   });
 
   test('should detect "All-or-Nothing" distortion', async () => {
-    // Mock sum to return a tensor with high similarity
-    const highSimTensor = {
-      dataSync: jest.fn(() => [0.95]),
-      dispose: jest.fn(),
-    };
-    tf.sum.mockReturnValue(highSimTensor);
-
-    mockEmbedFunction.mockResolvedValue({
-        squeeze: jest.fn(() => ({ dispose: jest.fn() })),
-        dispose: jest.fn(),
-        dataSync: jest.fn(() => [0.1]),
+    // Mock cosineDistance to return high similarity for the relevant CD type
+    // This requires knowing the order of CD types being checked in detectCDs
+    // For simplicity, we'll make all cosine distances high enough to trigger detection
+    // and then filter based on the expected outcome.
+    tf.metrics.cosineDistance.mockImplementation((vec1, vec2) => {
+      // In a real scenario, you might inspect vec1/vec2 to determine which CD is being checked
+      // For now, we'll just return a high similarity for all.
+      return createMockSqueezedTensor(0.95);
     });
 
     const text = 'I always fail at everything.';
-    const result = await detectCDs(text, 'medium');
-
-    expect(result.length).toBeGreaterThan(0);
+    const result = await detectCDs(text, 'high');
     expect(result).toContain('All-or-Nothing');
+    expect(use.load).toHaveBeenCalledTimes(1);
+    expect(mockEmbed).toHaveBeenCalled();
   });
 });
