+++ src/ui/SensitivitySelector.test.js
@@ -1,39 +1,35 @@
 import React from 'react';
-import { render, screen, within, fireEvent, act } from '@testing-library/react';
+import { render, screen, within } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
 import { SensitivitySelector } from './SensitivitySelector';
 
 describe('SensitivitySelector', () => {
   test('renders with initial value and options', async () => {
-    // Render the component
     render(<SensitivitySelector currentSensitivity="medium" onSensitivityChange={() => {}} />);
 
     // Check if the label is rendered
     expect(screen.getByLabelText('Sensitivity')).toBeInTheDocument();
 
-    // Check if the initial value is set correctly (MUI Select renders the value inside the combobox div)
-    // In some versions/configurations, MUI Select is a combobox, not a button
+    // Check if the initial value is set correctly
     expect(screen.getByRole('combobox')).toHaveTextContent('Medium');
+
+    // Check if all options are present (by opening the select)
+    await userEvent.click(screen.getByRole('combobox'));
+    const listbox = await screen.findByRole('listbox');
+    expect(listbox).toBeInTheDocument();
+    expect(within(listbox).getByText('Low')).toBeInTheDocument();
+    expect(within(listbox).getByText('Medium')).toBeInTheDocument();
+    expect(within(listbox).getByText('High')).toBeInTheDocument();
   });
 
   test('calls onSensitivityChange with the new value when changed', async () => {
     const handleChange = jest.fn();
     render(<SensitivitySelector currentSensitivity="medium" onSensitivityChange={handleChange} />);
 
-    // Open the select menu (MUI Select uses mouseDown to open)
-    // Use getByRole('combobox') as the trigger
-    const selectTrigger = screen.getByRole('combobox');
-    fireEvent.mouseDown(selectTrigger);
+    await userEvent.click(screen.getByRole('combobox'));
+    await userEvent.click(screen.getByText('High'));
 
-    // Wait for the listbox (menu options) to appear
-    const listbox = await screen.findByRole('listbox');
-    expect(listbox).toBeInTheDocument();
-
-    // Find and click the 'High' option inside the listbox
-    const highOption = within(listbox).getByText('High');
-    fireEvent.click(highOption);
-
-    // Assert that the callback was called with the correct value
     expect(handleChange).toHaveBeenCalledTimes(1);
     expect(handleChange).toHaveBeenCalledWith('high');
   });
